Bug 1279020 - [mach] Follow symlinked mach binary when searching for a srcdir, r=gps
The mach binary needs to find a srcdir (or test zip) to operator on. As a last resort,
it checks the directory containing the binary itself. If the binary has been symlinked
to say, $HOME/bin, it won't follow the symlink to the proper location, and the check
will fail.

Fixing this means developers can run `mach` from outside of the srcdir, as the symlink
will be followed to the actual srcdir. The motivation for this is actually to improve
the UX of debugging failures on interactive workers. With this fix, we can provide
developers with a mach environment directly from $HOME without them needing to go
find it.

MozReview-Commit-ID: ZEH8MuyKv

UltraBlame original commit: 80eb5947da076120f095ac7dd6a32670032e0134