Bug 1266456 - part10: fix tests for autocomplete-popup HTMLTooltip migration;r=bgrins
Most test modifications deal with listening to the proper events.
As mentioned in the parent commit, the keyboard navigation for the
inspector-search changed a bit with this change, so the related test
case is using a slightly different scenario.

MozReview-Commit-ID: BgSa69bSSFB

UltraBlame original commit: 1de829f2f1f03e23ca0159bee473d36b9989e62b