Bug 1273725 - Don't show 'No supported source' message in video controls if there's no sources. r=jaws
We're flashing a warning message  "NO video with supported format and MIME type
found." when we're loading a YouTube page in Fennec. This is because YouTube
are calling load() on a video element with no src attribute or source children
in a click handler in order to defeat the "requires user interaction to play"
restrictions on mobile Chrome and Safari in that video element.

We're showing the warning message to help developers diagnose why loads are
unintentionally failing. However, the behaviour that YouTube are doing is
intentional, and YouTube are likely to not be the only ones doing this. So we
should change our video controls to not display the warning message in this
case. We should be compatible with other browsers.

We'll still log a debug message to the browser console as to why the load
is failing, to help developers who hit this unintentionally.

MozReview-Commit-ID: IpugbMOloJt

UltraBlame original commit: afb94913764c661534e600a8ce4ad14b5289221c