Bug 1278890 - Add a method for normalizing test paths to the test package mach environment, r=armenzg
The test path structure is slightly different in the test package compared to a srcdir. So we may need
to normalize the specified paths such that they are relative to a srcdir. Because every test harness
needs to do this, this method is being added to the bootstrap for re-use.

MozReview-Commit-ID: fBAbfuG5XQ

UltraBlame original commit: 45d6a990dd2877b9c8f265f13fc565a96f6667bf