Bug 1287042 - Always before/after CSS rules in rule-view; r=bgrins
::before and ::after pseudo-elements are visible in the markup-view today
but if, for some reason, they aren't generated, we still want to know that
the CSS rule exists.
This may happen if you use display:none on the pseudo-element CSS rule itself.
When that happens, the pseudo-element won't be generated and therefore there
will be no possibility to see the rule in the rule-view (you'd have to go to
the style-editor for that).
This change keeps the pseudo-elements in the markup-view, but also adds the
corresponding CSS rules in the rule-view.

MozReview-Commit-ID: tx5IpmtE7b

UltraBlame original commit: ec882720349ca69e63f2c1ca7b5bff6c30ff18e2