Bug 1247168 - Change decision image to Ubuntu 16.04; r=dustin
We're currently running Ubuntu 14.04 in the decision image. While it is
still in LTS support, 16.04 ships with a modern, properly configured Python
2.7. So we upgrade to 16.04 and drop the install of Python from source
because it is no longer needed.

This is part 1 of a larger refactor to this image.

MozReview-Commit-ID: CTbsPmTjcgs

UltraBlame original commit: eb08ab6d23493e62fecaa2b2d7ec44119a34ff6c