Bug 1247168 - Actually use workspace cache in decision task; r=dustin
The decision task configures a /home/worker/workspace cache. However,
the command we run in the container references a "workspace" relative
path. From logs in automation, it appears that PWD during execution
is "/" because "workspace" is being resolved to "/workspace." The
net result of this is we appear to be performing a VCS clone+checkout
on every single task.

This commit fixes the paths so our workspace cache is actually used.

MozReview-Commit-ID: Kj6REep5bSs

UltraBlame original commit: 6c9ced9abf7eae19b05bf10420770dce0b20c045