Bug 1289550 - Remove "fun little hack" from XPCNativeSetKey. r=bz
XPCNativeSetKey has a huge comment about this weird hack it does,
where it tags the first 16 bytes with a magic value. The purpose of
this seem to be that PLDHashtable used to require that the Match()
operation handle both the desired "key" type and the actual entry type
(NativeSetMap::Entry in this case), with the latter needed for
resizing. However, that duality in the match operation has not been
needed since bug 374906, which landed in 2007, so this class can be
greatly simplified.

IsAKey() can be replaced with true, which simplifies some hash
operations.

UltraBlame original commit: 8830f07a659b22aa291704ff0eb8c42fd7368871