Bug 1289678 - Don't count audio stream creation failures when retrying on Telemetry. r=kinetik
When failing to create an audio stream, we fallback to a SystemClockDriver
marked as being a "fallback driver". When failing again to open an audio stream
after re-trying, we can check whether we came from a fallback driver, and not
report the failure again to telemetry.

MozReview-Commit-ID: FAdQ0pCtC3m

UltraBlame original commit: 58f01ab6d8560966453ee42d2f1f07c02c99b1e3