Bug 1290019 - Remove MT version check. r=chmanchester
Before bug 1289294, the impossibility to find the version of MT was only
issuing a warning. Warnings in configure are essentially useless, so
since we weren't and still aren't doing anything with the result of that
check, and since there are versions of MT that don't print out a version
number, just remove the check.

UltraBlame original commit: b722fac1af4b88f9ed41f620b0eef2e851f04faa