Bug 1168674: [vorbis] P6. Do not tweak first decoded sample time. r=cpearce r?cpearce
A hack was in place to ensure that the first returned sample would have
a time of 0 so that the media start time would be 0. This was incorrect
for two primary reasons:
- The media start time is adjusted according to the first sample anyway.
- When seeking to 0, the first sample would have a time different to the
first sample decoded (when we want them to have the same time).

MozReview-Commit-ID: IyuT9O2F4EZ

UltraBlame original commit: 864104ba08a648137660176cca3618ec183c31e8