Bug 1290531 - Remove tester-device Docker image; r=dustin
There are no references to tester-device in tree or in the gaia repo.
Since it appears to be unused, remove it.

According to garndt, this was used for testing in the remote device
lab, which he things is no longer happening.

The reason I found this is because it is the only thing using a custom
"build.sh" to create Docker images. I'm rewriting the Docker image
building functionality and tester-device is a one-off interfering
with that work. Making it go away is the easiest way to unblock me.

MozReview-Commit-ID: 4veem4uIXXi

UltraBlame original commit: 1532b022e19cf3f6b483649fb6dccbbae7a5226f