Bug 1291811 - Move GeckoView Java sources into mobile/android/geckoview. r=sebastian
This is the first step only.  It moves Experiments out of util (just
for convenience) -- I would have liked to have moved WhatsNewReceiver
out of notifications as well, but it's referenced in
AndroidManifest.xml.in and I don't want to break that.

The set of files in GeckoView is slightly larger than what we really
want, but it's close to correct.  Future tickets will get things
compiling and trim the set of files down to size.  Hopefully we can
move notifications/, permissions/, and restrictions/ back to Fennec.

This is not technically required, but it's best to bulk move the files
up front and patch later, rather than jump through hoops to
distinguish the set of files in Gradle (and moz.build).

MozReview-Commit-ID: EWaAeswFoeG

UltraBlame original commit: f47c2d2a6e27cc0b8f09650c97eb7bd8c72fc66a