Bug 1201363 - MediaStreamVideoSink for ImageCapture case. r=jesup
Make CaptureTask to inherite from MediaStreamVideoSink. The main change is to move the logic of |NotifyQueuedTrackChanges| to |SetCurrentFrames|.
The original image capture is not modified for support multiple video MediaStreamTracks. The design still used the track id in owned media stream. The should be fixed in the following bug if we still want to support ImageCapture in multiple video tracks case.

MozReview-Commit-ID: Od4tHoR8Ef

UltraBlame original commit: 7c07c4aa3569689c3062557ed1115cbc3f5ffdbb