Bug 1292696 - Pin various graphics related packages to known good versions; r=jmaher
Without this pinning, the image doesn't build deterministically and
fails the version checks that were previously in this file. At the
time this was written, the libgl1-mesa packages were on version
11.2.0-1ubuntu2.1 upstream and building the Docker image locally
failed because these versions would get installed.

A risk to pinning is we'll drift out of date with the OS we're testing
and won't detect failures on packages users in the wild are running.
But on the flip side, automation will be able to build the image
reliably and there won't be "random" failures due to automation
picking up new package versions.

MozReview-Commit-ID: 5MSEbsiNjF1

UltraBlame original commit: 0809927f1e16dd46cc4c8a29c5a46c31a3cfc1e3