Bug 1293211 - part2: fire ready event in SwatchBasedEditorTooltips after widget initialization;r=bgrins
Now that showing a XUL panel is asynchronous, tests need to be able to wait for the complete
initialization of the editor tooltips. Waiting for the tooltip 'shown' event is not enough
here because tooltips are also waiting for this event to start initializing their widgets.

MozReview-Commit-ID: DGTyeVrHNb

UltraBlame original commit: a98ce7344c9a7192a0127c7e7354500f2ddc288b