Bug 1292278 - Zero-index the namespace manager URIs. r=bz
The current setup here tries to "save" an entry in the list by checking
for the "None" namespace everywhere and special-casing it. We can simplify
this a lot by just adding the empty atom to the beginning of the array,
which is the value that servo uses to represent "namespace none" anyway.

UltraBlame original commit: 2629e4786dd8d07169068f86b84d7c24670306fd